Adding subnet allocation type "any"
This commit adds an option for subnet allocation type "any".
In the current azure ipam code base, if the Subnet option is
passed to the ipam plugin, the ipam plugin tries to allocate
from within the subnet pool if it is availabled. If the subnet
option is missing, it requests a pool if its available and
allocated IP from that pool. However there is no option to
allocate an IP from any available subnet pool.
This commits adds a logic Subnet is set to 0.0.0.0/24 then the
azure ipam plugin will allocate an IP from any subnet pool
available on the interface. This will avoid the need to have
local configuration on every node in Azure when the nodes
are connected to different subpools in the vnet.

Signed-off-by: Abhinandan Prativadi <abhi@docker.com>