Add HomeAZ to NodeInfoStatus (#3738)
In service of adding HomeAZ awareness to apiserver-vnet-integration,
this adds HomeAZ to the NodeInfoStatus portion of the NodeInfo CRD.
Similar to other elements of NodeInfoStatus, HomeAZ is an optional
property of a node that can be communicated. If present, it strictly
follows the AZXX format, where XX are two 0-9 integers.