Move HomeAZ to NodeInfo Spec
It was discovered that HomeAZ was incorrectly placed in the Status
portion of NodeInfo, rather than the Spec, where it should be placed.
This is because information in Spec is provided by CNS, while
information in the Status is provided by DNC. As CNS is the entity that
is initially aware of the HomeAZ of some node, the correct place for
this to be communicated is in the Spec. Comments have also been updated
to reflect the ownership of these fields, so as to not cause the same
confusion to future contributors.